Code cleanup
- Use new SequencedCollection getters
- Apply `abstract` to all mixins
- Add private constructor to all Util classes

+++ src/main/java/meteordevelopment/meteorclient/systems/modules/render/Freecam.java
@@ -168,7 +168,7 @@ public class Freecam extends Module {
         if (reloadChunks.get()) mc.worldRenderer.reload();
         mc.options.setPerspective(perspective);
         if (staticView.get()) {
-            mc.options.getFovEffectScale().setValue((double)fovScale);
+            mc.options.getFovEffectScale().setValue(fovScale);
             mc.options.getBobView().setValue(bobView);
         }
     }
